– Dec 9, 24 What we have: – server and client have been coded but the client needs a little work –main menu is mainly done –map format parser needs to be made – splash text is ready to be added to the main menu – skeleton of the movement is done just needs to be connected to the player png and map – update system is no longer needed due to the hypersonic games discord server turning to a community server –skeleton of the settings manager is done (and by that i mean the code registers input)
What we need:
–weapons system
–inventory system
–mods (very last thing we need)
–if we have not coded a way for players to join a hosted game we need to do that
– a website to be able to access the game

What is being worked on/ talked about
            –supabase integration for updates/leaderboards
            – using the SVG format to make the logo for the main menu
  – making voice lines for stuff like player count call outs, opening voiceline when opening the game or when you start the tutorial etc
–  (Yaamee) Working on coding main menu buttons.

– April 2, 25
What needs to be done: weapons manager (i think?), server logic needs to be finished up and client logic needs to be implemented into server.py, client.py needs to be made but i'm pretty sure that needs to wait on server.py’s client logic to begin work on that, i think we also need to implement the map system but im not to sure if its even made, death logic needs to be made, movement logic needs to be made, maybe an inventory system but that's a BIG maybe because if we go with a certain route aka where you spawn with a different weapon each time you respawn we might not even need it, possibly a way for people to join hosted games but we can do that once we get server and client setup, player callouts, and i think the final thing we need is a settings manager, possibly mods

What we have: main menu is pretty much done just needs button input and logo, we have pretty much every asset, death sound (not yet implemented in the game), credit and main menu music, intro voice line when you first boot up the game (in the files but not yet implemented into the actual game)

ideas/ questions and anything of note: example format: Q: (name) - [question or note here]
                                                                                        A: (name) - [awnser here]
Q: (everett): hey sebastian are you still working on that python bytecode hook/patch system for mods?
A:

Q: (everett): Yaamee are you still workin on finding a way to code the main menu buttons?
A:

Note: (everett) i could possibly do the player callout voice lines but if anyone else would like to do it please let me know

– April 4, 25
What needs to be done: weapons manager (i think?), server logic needs to be finished up and client logic needs to be implemented into server.py, client.py needs to be made but i'm pretty sure that needs to wait on server.py’s client logic to begin work on that, i think we also need to implement the map system but im not to sure if its even made, death logic needs to be made, movement logic needs to be made, maybe an inventory system but that's a BIG maybe because if we go with a certain route aka where you spawn with a different weapon each time you respawn we might not even need it, possibly a way for people to join hosted games but we can do that once we get server and client setup, player callouts, and i think the final thing we need is a settings manager, possibly mods

What we have: main menu is pretty much done just needs button input and logo, we have pretty much every asset, death sound (not yet implemented in the game), credit and main menu music, intro voice line when you first boot up the game (in the files but not yet implemented into the actual game)

ideas/ questions and anything of note: example format: Q: (name) - [question or note here]
                                                                                        A: (name) - [awnser here]

Q: (everett): hey sebastian are you still working on that python bytecode hook/patch system for mods?
A:

Q: (everett): Yaamee are you still workin on finding a way to code the main menu buttons?
A:

Note: (everett) i could possibly do the player callout voice lines but if anyone else would like to do it please let me know

Note: (everett) i think i found some code we could work with to make the map and implement it into the server
Code: class Map():
  def __init__(self):
    self.Image = pygame.image.load("testbackground.jpg").convert()
    self.rect = self.Image.get_rect()
    self.rect.x = 0
    self.rect.y = 0
This code we can start with then make a map.txt and put that in the code then we can go from there
Q: (everett) btw hey Sebastain do you know if we even have a map.txt?

-apr 6, 2025
What needs to be done:, server logic needs to be finished up and client logic needs to be implemented into server.py, client.py needs to be made but i'm pretty sure that needs to wait on server.py’s client logic to begin work on that, i think we also need to implement the map system but im not to sure if its even made, death logic needs to be made, movement logic needs to be made, maybe an inventory system but that's a BIG maybe because if we go with a certain route aka where you spawn with a different weapon each time you respawn we might not even need it, possibly a way for people to join hosted games but we can do that once we get server and client setup, player callouts, possibly mods,

What we have: main menu is pretty much done just needs button input and logo, we have pretty much every asset, death sound (not yet implemented in the game), credit and main menu music, intro voice line when you first boot up the game (in the files but not yet implemented into the actual game), map txt (not yet implemented into the game but it is in the file, weapons manager has code but i am unsure if its complete, settings manager also has code but im unsure if its complete too

ideas/ questions and anything of note: example format: Q: (name) - [question or note here]
                                                                                        A: (name) - [awnser here]

Q: (everett): hey sebastian are you still working on that python bytecode hook/patch system for mods?
A:

Q: (everett): Yaamee are you still workin on finding a way to code the main menu buttons?
A:

Note: (everett) i could possibly do the player callout voice lines but if anyone else would like to do it please let me know

Note: (everett) i think i found some code we could work with to make the map and implement it into the server
Code: class Map():
  def __init__(self):
    self.Image = pygame.image.load("testbackground.jpg").convert()
    self.rect = self.Image.get_rect()
    self.rect.x = 0
    self.rect.y = 0
This code we can start with then make a map.txt and put that in the code then we can go from there